글의 유무 체

+++ src/main/java/org/example/App.java
@@ -115,9 +115,19 @@ public class App {
                 System.out.println("번호는 정수로 입력해");
                 return 0;
             }
-            SecSql sql = new SecSql();
 
-            DBUtil.selectRowIntValue(conn, sql);
+            SecSql sql = new SecSql();
+            sql.append("Select *");
+            sql.append("FROM article");
+            sql.append("WHERE id = ?;", id);
+
+            Map<String, Object> selectRow = DBUtil.selectRow(conn, sql);
+
+            if (selectRow.isEmpty()) {
+                System.out.println(id + "번 게시물은 없습니다.");
+                return 0;
+            }
+
 
             System.out.println("==수정==");
             System.out.print("새 제목 : ");
@@ -125,6 +135,7 @@ public class App {
             System.out.print("새 내용 : ");
             String body = sc.nextLine().trim();
 
+            sql = new SecSql();
             sql.append("UPDATE article");
             sql.append("SET updateDate = NOW()");
             if (!title.isEmpty()) { // title 값이 들어있을 때
@@ -149,9 +160,21 @@ public class App {
                 return 0;
             }
 
+            SecSql sql = new SecSql();
+            sql.append("Select *");
+            sql.append("FROM article");
+            sql.append("WHERE id = ?;", id);
+
+            Map<String, Object> selectRow = DBUtil.selectRow(conn, sql);
+
+            if (selectRow.isEmpty()) {
+                System.out.println(id + "번 게시물은 없습니다.");
+                return 0;
+            }
+
             Article foundArticle = null;
 
-            SecSql sql = new SecSql();
+            sql = new SecSql();
             sql.append("SELECT *");
             sql.append("FROM article");
             sql.append("WHERE id = ?;", id);
@@ -185,6 +208,18 @@ public class App {
             }
 
             SecSql sql = new SecSql();
+            sql.append("Select *");
+            sql.append("FROM article");
+            sql.append("WHERE id = ?;", id);
+
+            Map<String, Object> selectRow = DBUtil.selectRow(conn, sql);
+
+            if (selectRow.isEmpty()) {
+                System.out.println(id + "번 게시물은 없습니다.");
+                return 0;
+            }
+
+            sql = new SecSql();
             sql.append("DELETE FROM article");
             sql.append("WHERE id = ?;", id);
